Make the use of apxs/apxs2 consistent in the configure script.
The configure script advertised a --with-apxs option. That option did
what it was supposed to, except that the configure script would
complain that --with-apxs was not a supported option. The problem was
that the argument was defined as --with-apxs2 in one spot, but used as
--with-apxs elsewhere.

This change resolves the naming conflict in ./configure in favor of
--with-apxs rather than --with-apxs2.