Abstract out the building of cas_saml_attr lists.
There are currently three places in mod_auth_cas that build
cas_saml_attr linked lists. Those places had a fair amount of
duplicated code, that was relatively error-prone (pointer munging).
This patch creates an abstraction for building lists of cas_saml_attrs
with a simple API. Use of that API ensures that memory is managed
consistently and correctly and attribute structures are consistent
regardless of how they were created. It also makes it easier to test
these code paths.