Fix the argument passed to ap_get/set_module_config in cas_get/set_attributes
The parameter was incorrectly set to the request_rec pointer itself
rather than the request_config field of the request. Depending on
Apache's configuration, this resulted in different parts of the
request structure being overwritten with the attribute pointer. In
many cases, it worked correctly by accident. This change correctly
uses the request_config conf_vector for ap_get/set_module_config.

In addition to that change, cas_get_attributes also works better in a
subrequest: if no attributes are found in the subrequest, it will look
in the parent request for attributes.