Use ap_hook_check_authn() instead of ap_hook_check_access().
This allows "Satisy Any" and the 2.4-style multiple Require(Any|All)
lines to work properly.

The Apache 2.4 docs say the following about ap_hook_check_authn():

This hook is used to analyze the request headers, authenticate the user,
and set the user information in the request record (r->user and
r->ap_auth_type). This hook is only run when Apache determines that
authentication/authorization is required for this resource (as
determined by the 'Require' directive). It runs after the access_checker
hook, and before the auth_checker hook. This hook should be registered
with ap_hook_check_authn(). If "Satisfy any" is in effect, this hook may
be skipped.

AUTHZ_DENIED_NO_USER needs to be returned in cas_check_authorization()
in case the user hasn't been authenticated yet.

Add stub for ap_hook_check_authn().