Remove CAS_MAX_RESPONSE_SIZE.
Use a dyanamic buffer to capture all of the response from the CAS
server.

This strategy is outlined in the example refereced here:

https://curl.haxx.se/libcurl/c/CURLOPT_WRITEFUNCTION.html

APR doesn't have a realloc(), so we call apr_pcalloc() on every call to
cas_curl_write(). This is wasteful. Should we create a separate memory
pool to handle this so we can free memory that is no longer used?