Use a dynamic buffer to store the CAS validation response. (#132)
Use a dynamic buffer to capture all of the response from the CAS
server.

This strategy is outlined in the example referenced here:

https://curl.haxx.se/libcurl/c/CURLOPT_WRITEFUNCTION.html

APR doesn't have a realloc(), so we create a subpool on every call to
cas_curl_write() to allocate memory from. Previous subpools are
destroyed to prevent memory waste.

CAS_MAX_RESPONSE_SIZE has been increased to a large value to
prevent users from having to modify it for large responses.