Replace bespoke cookie parsing with ap_cookie_read() (#202)
* Replace bespoke cookie parsing with ap_cookie_read()

* Fix escaping for small chars in urlEncode()

The %%%x format string resolves to the literal "%" and the hex representation
of the character to be encoded, but is always asssumed to return three characters.

However for a small value like e.g. 7 it would return "%7" instead. None of the
current two call sites of the function use such a small value, but apply correct
padding just in case the function might be used elsewhere in the future.

* Update src/mod_auth_cas.c

Co-authored-by: David Hawes <dhawes@gmail.com>

* Update docs to require Apache 2.4

The upstream support for Apache 2.2.x ended on 2018-01-01 and also none of
the long term Linux distros still support it, looking at the latest still
supported releases:

* Debian 8 ELTS has Apache httpd 2.4.10
* Ubuntu 14.4 has Apache httpd 2.4.5
* RHEL 7 has Apache httpd 2.4.6
* SLES 11 has Apache httpd 2.4.23

Co-authored-by: Moritz Muehlenhoff <jmm@debian.org>
Co-authored-by: Moritz Muehlenhoff <moritz@wikimedia.org>
Co-authored-by: David Hawes <dhawes@gmail.com>