Revert ap_cookie_read() cookie parsing (#207)
ap_cookie_read() returns an error if there are duplicated cookies, which
may cause issues for current mod_auth_cas configurations.
ap_cookie_read() uses apr_table_do() with a static function to extract
the cookie line, so it can't be used directly. Revert to the original
cookie parsing code for the time being.

Prevent reading past the end of the cookie.